+++ python/helpers/lib/parser.py
@@ -168,29 +168,36 @@ def parse_setup_cfg(directory):
             for req in requires:
                 parse_requirement(req, req_type)
 
-        config = configparser.ConfigParser()
-        config.read(directory + "/setup.cfg")
+        try:
+            config = configparser.ConfigParser()
+            # preserve case
+            config.optionxform = str
+            config.read(directory + "/setup.cfg")
 
-        for req_type in [
-            "setup_requires",
-            "install_requires",
-            "tests_require",
-        ]:
-            requires = [
-                r
-                for r in config.get("options", req_type, fallback="")
-                .strip()
-                .split("\n")
-                if r != ""
-            ]
-            parse_requirements(requires, req_type)
+            for req_type in [
+                "setup_requires",
+                "install_requires",
+                "tests_require",
+            ]:
+                requires = [
+                    r
+                    for r in config.get("options", req_type, fallback="")
+                    .strip()
+                    .split("\n")
+                    if r != ""
+                ]
+                parse_requirements(requires, req_type)
 
-        extras_require_section = "options.extras_require"
-        if config.has_section(extras_require_section):
-            section = config[extras_require_section]
-            for key, value in section.items():
-                parse_requirements(
-                    value.strip().split("\n"), "extras_require:{}".format(key)
-                )
+                extras_require_section = "options.extras_require"
+                if config.has_section(extras_require_section):
+                    section = config[extras_require_section]
+                    for key, value in section.items():
+                        parse_requirements(
+                            value.strip().split("\n"),
+                            "extras_require:{}".format(key),
+                        )
+        except Exception as e:
+            print(json.dumps({"error": repr(e)}))
+            exit(1)
 
     return json.dumps({"result": setup_packages})
