+++ python/lib/dependabot/python/file_fetcher.rb
@@ -47,7 +47,7 @@ module Dependabot
         fetched_files += requirement_files if requirements_txt_files.any?
 
         fetched_files << setup_file if setup_file
-        fetched_files << setup_cfg if setup_cfg
+        fetched_files << setup_cfg_file if setup_cfg_file
         fetched_files += path_setup_files
         fetched_files << pip_conf if pip_conf
         fetched_files << python_version if python_version
@@ -79,7 +79,7 @@ module Dependabot
       end
 
       def check_required_files_present
-        return if requirements_txt_files.any? || setup_file || pipfile || pyproject
+        return if requirements_txt_files.any? || setup_file || setup_cfg_file || pipfile || pyproject
 
         path = Pathname.new(File.join(directory, "requirements.txt")).
                cleanpath.to_path
@@ -90,8 +90,8 @@ module Dependabot
         @setup_file ||= fetch_file_if_present("setup.py")
       end
 
-      def setup_cfg
-        @setup_cfg ||= fetch_file_if_present("setup.cfg")
+      def setup_cfg_file
+        @setup_cfg_file ||= fetch_file_if_present("setup.cfg")
       end
 
       def pip_conf
