+++ python/spec/dependabot/python/file_parser/setup_cfg_file_parser_spec.rb
@@ -110,7 +110,7 @@ RSpec.describe Dependabot::Python::FileParser::SetupCfgFileParser do
             expect(error.class).
               to eq(Dependabot::DependencyFileNotEvaluatable)
             expect(error.message).
-              to eq('Illformed requirement ["==2.6.1raven==5.32.0"]')
+              to eq("InstallationError(\"Invalid requirement: 'psycopg2==2.6.1raven == 5.32.0'\")")
           end
       end
     end
