+++ python/spec/fixtures/setup_files/illformed_req.cfg
@@ -1,3 +1,4 @@
+[metadata]
 name = python-package
 version = 0.0
 description = Example setup.cfg
