+++ python/spec/fixtures/setup_files/markers.cfg
@@ -1,3 +1,4 @@
+[metadata]
 name = python-package
 version = 0.0
 description = Example setup.cfg
