+++ python/spec/fixtures/setup_files/no_tests_require.cfg
@@ -1,3 +1,4 @@
+[metadata]
 name = python-package
 version = 0.0
 description = Example setup.cfg
